Update build script
Updated appveyor.yml, the build script has been extracted to "appveyor /
build.cmd", facilitating its maintenance.
In build.cmd loops are used to optimize the construction of the
different versions:

NTS-x64
TS-x64
NTS-x86
TS-x86
In addition, all versions of the extension are packaged in a single zip
file.
The php example scripts have also been moved to the examples folder.
This folder is included with the zip file of the extension binaries.